More improvement on vehicle controller
Made the ball smaller.
Made the ball behave less floaty.
Only accelerates when touching the ground.
Added a Hop function. (Planning it to proceed to drift when held)
Added a Boost function. (Right now it only increases the max speed)

Did a little mapping to make it look nicer, and to practice some technique.