Moved unused scripts under the rug
Spent some time figuring out how exactly these "starter assets" are set up to work.
Moved a bunch of crap around so it's less in the way. I plan to take a lot of the content from the initial folder to reduce load times an clutter on my end.

I also experimented with making the player a child to the vehicle.